backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# WebSocket
websockets==12.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS & Middleware
python-dotenv==1.0.0

# File Processing
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1

# Date & Time
python-dateutil==2.8.2

# CSV & Excel
pandas==2.1.3
openpyxl==3.1.2

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
python-json-logger==2.0.7

# Integration with existing scripts
boto3==1.29.7  # For AWS Textract
openai==1.3.7  # For GPT integration
